spring-2: refactor codes based on feedback from metor Scott.
