+++ src/main/java/com/safetynet/alerts/controller/MedicalRecordController.java
@@ -1,12 +1,9 @@
 package com.safetynet.alerts.controller;
 
 import com.safetynet.alerts.model.MedicalRecord;
-import com.safetynet.alerts.service.MedicalRecordServiceImpl;
+import com.safetynet.alerts.service.MedicalRecordService;
 
-import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,7 +14,7 @@ import java.util.List;
 public class MedicalRecordController {
 
     @Autowired
-    MedicalRecordServiceImpl medicalRecordService;
+    MedicalRecordService medicalRecordService;
 
     @GetMapping
     public ResponseEntity<List<MedicalRecord>> getALlMedicalRecords(){
