+++ src/main/java/com/safetynet/alerts/service/FireStationService.java
@@ -1,8 +1,19 @@
 package com.safetynet.alerts.service;
 
-
 import com.safetynet.alerts.model.FireStation;
+import com.safetynet.alerts.repository.FireStationRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
-public interface FireStationService {
-    FireStation saveFireStation(FireStation fireStation);
+@Slf4j
+@Service
+public class FireStationService {
+
+    @Autowired
+    FireStationRepository fireStationRepository;
+
+    public FireStation saveFireStation(FireStation fireStation) {
+        return fireStationRepository.save(fireStation);
+    }
 }
