--- src/main/java/com/safetynet/alerts/service/FireStationServiceImpl.java
@@ -1,20 +1,0 @@
-package com.safetynet.alerts.service;
-
-import com.safetynet.alerts.model.FireStation;
-import com.safetynet.alerts.repository.FireStationRepository;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Slf4j
-@Service
-public class FireStationServiceImpl implements FireStationService{
-
-    @Autowired
-    FireStationRepository fireStationRepository;
-
-    @Override
-    public FireStation saveFireStation(FireStation fireStation){
-        return fireStationRepository.save(fireStation);
-    }
-}
