+++ src/main/java/com/safetynet/alerts/service/MedicalRecordService.java
@@ -1,12 +1,24 @@
 package com.safetynet.alerts.service;
 
 import com.safetynet.alerts.model.MedicalRecord;
+import com.safetynet.alerts.repository.MedicalRecordRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-public interface MedicalRecordService {
-    ResponseEntity<List<MedicalRecord>> getALlMedicalRecords();
-//    ResponseEntity<HttpStatus> addNewMedicalRecord(MedicalRecord medicalRecord);
+@Slf4j
+@Service
+public class MedicalRecordService {
+
+    @Autowired
+    private MedicalRecordRepository medicalRecordRepository;
+
+    public ResponseEntity<List<MedicalRecord>> getALlMedicalRecords() {
+        return new ResponseEntity<>(medicalRecordRepository.findAll(), HttpStatus.FOUND);
+    }
+
 }
