--- src/main/java/com/safetynet/alerts/service/MedicalRecordServiceImpl.java
@@ -1,26 +1,0 @@
-package com.safetynet.alerts.service;
-
-import com.safetynet.alerts.model.MedicalRecord;
-import com.safetynet.alerts.repository.MedicalRecordRepository;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Slf4j
-@Service
-public class MedicalRecordServiceImpl implements MedicalRecordService {
-
-    @Autowired
-    private MedicalRecordRepository medicalRecordRepository;
-
-    @Override
-    public ResponseEntity<List<MedicalRecord>> getALlMedicalRecords() {
-        return new ResponseEntity<>(medicalRecordRepository.findAll(), HttpStatus.FOUND);
-    }
-
-}
