+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
 spring.application.name=SafetyNet Alerts Backend
+management.endpoints.web.exposure.include=health,info,metrics,httpexchanges
